refactor(welcomeScreen): add custom hook for story markers
