+++ src/scripts/components/splash-screen/splash-screen.styl
@@ -21,7 +21,7 @@
     text-transform: uppercase
     font-weight: bold
     font-size: emCalc(36px)
-    font-family: NotesEsa
+    font-family: NotesEsaBold
     line-height: emCalc(16px)
 
   p
